api/app/requirements.txt
wheel
requests
loguru

uvicorn
websockets
pydantic
fastapi
fastapi-cache2[redis]
fastapi-socketio
fastapi-utils

python-keycloak
